Add isomiR analysis
Add the ability to analyze isomiRs along with canonical miRNAs. A file containing isomiR counts, sequences, and RPM is output when the --isomir flag is used. The flag also causes the alignment percentages to be split into sequences that map either perfectly or with 1 bp mismatch to standard miRNAs. A HashMap is used for sequence comparison rather than traditional alignment for both speed and to make sure truncated sequences won't match. When not checking for isomiRs, a simple bowtie2 alignment is used. While this does work pretty well, it considers 3' truncated miRNAs to align just as well as full length products. There may be a setting to disallow this, but I'm not certain what it is. This leads to an overestimation of the number of unique miRNAs and their counts. Other changes include documentation updates, making the code more idiomatic, adding in a max length variable, and other minor changes.